+++ packages/webview_flutter/example/lib/main.dart
@@ -75,6 +75,7 @@ class _WebViewExampleState extends State<WebViewExample> {
             print('Page finished loading: $url');
           },
           gestureNavigationEnabled: true,
+          useExperimentalAndroidSurfaceView: true,
         );
       }),
       floatingActionButton: favoriteButton(),